Move previous demo to PreviousDemo component, created Results and TenantForm
