+++ client/src/components/Demo/Results.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export const Results = () => {
+    return (
+        <div>
+            <h2>Results</h2>
+            <ul>
+                <li>Tenant 1</li>
+                <li>Tenant 2</li>
+                <li>Tenant 3</li>
+            </ul>
+        </div>
+    );
+};
