+++ client/src/components/Demo/index.js
@@ -1,0 +1,3 @@
+export { TenantForm } from './TenantForm';
+export { PreviousDemo } from './PreviousDemo';
+export { Results } from './Results';
